Update dependencies
scButterfly is used as a submodule for the main repo of the thesis. I have attempted to resolve all the dependencies using conda (thesis repo), and pip install only for the absolute necessarry pip only available packages, and the package itself (pip install -e .).

scvi-colab isn't available with conda, and not necessarily for our study as well. So it was removed.